build: moved ciphers in subdirectories
