+++ .gitignore
@@ -52,3 +52,4 @@ Mkfile.old
 dkms.conf
 .vscode/
 Makefile
+Makefile
